+++ src/locales/hu/sidebar.ts
@@ -38,6 +38,7 @@ export const sidebar = {
   libraryDeselect: '{{name}} kizárása',
   librarySelectionCount: '{{count}} könyvtár',
   serverSelectionCount: '{{count}} szerver',
+  serverAvailabilityCount: '<unavailable>{{total}}</unavailable>→{{available}} szerver',
   expandPlaylists: 'Lejátszási listák kibontása',
   collapsePlaylists: 'Lejátszási listák összecsukása',
   more: 'Több',