+++ src/locales/nb/sidebar.ts
@@ -34,6 +34,7 @@ export const sidebar = {
   libraryDeselect: 'Ekskluder {{name}}',
   librarySelectionCount: '{{count}} biblioteker',
   serverSelectionCount: '{{count}} servere',
+  serverAvailabilityCount: '<unavailable>{{total}}</unavailable>→{{available}} servere',
   expandPlaylists: 'Utvid spillelister',
   collapsePlaylists: 'Skjul spillelister',
   more: 'Mer',